+++ src/api/questions.js
@@ -13,7 +13,7 @@ const {
   requireQueueForQuestion,
   requireQuestion,
   failIfErrors,
-  canUserSeeQuestionDetailsForConfidentialQueue,
+  isUserStudent,
   filterConfidentialQueueQuestionsForUser,
   ApiError,
 } = require('./util')
@@ -137,17 +137,21 @@ router.post(
       }
     }
 
-    const question = Question.build({
-      name: data.name,
-      // Questions in fixed-location queues should never have a location
-      location: res.locals.queue.fixedLocation ? '' : data.location,
-      topic: data.topic,
-      enqueueTime: new Date(),
-      queueId,
-      askedById: askerId,
-    })
+    const question = await Question.create(
+      {
+        name: data.name,
+        // Questions in fixed-location queues should never have a location
+        location: res.locals.queue.fixedLocation ? '' : data.location,
+        topic: data.topic,
+        enqueueTime: new Date(),
+        queueId,
+        askedById: askerId,
+      },
+      {
+        include: [{ model: User, as: 'askedBy' }],
+      }
+    )
 
-    await question.save()
     await question.reload()
     res.status(201).send(question)
   })
@@ -173,7 +177,7 @@ router.get(
     )
     if (isConfidential) {
       const { userAuthz } = res.locals
-      if (!canUserSeeQuestionDetailsForConfidentialQueue(userAuthz, courseId)) {
+      if (isUserStudent(userAuthz, courseId)) {
         const { id: userId } = res.locals.userAuthn
         questions = filterConfidentialQueueQuestionsForUser(userId, questions)
       }
@@ -191,7 +195,7 @@ router.get(
     if (isConfidential) {
       const { id: userId } = res.locals.userAuthn
       const { userAuthz } = res.locals
-      if (!canUserSeeQuestionDetailsForConfidentialQueue(userAuthz, courseId)) {
+      if (isUserStudent(userAuthz, courseId)) {
         if (res.locals.question.askedById !== userId) {
           res.status(403).send('You are not authorized to access that question')
           return